requirements.txt
igdb_api_v4==0.3.2
python-dotenv==1.0.1
Requests==2.31.0
